Add mechanism for sensor-instantiation retry
It's a little ugly, but for various reasons sometimes things don't
always work on the first attempt and we need to try again a moment
later.

For now we hard-code a maximum of 5 attempts with a delay of 3 seconds
between each; if needed these parameters can be further refined later.