threshold: Only assert alarms on readings & thresholds with valid values
If either the sample or the threshold value is bogus (e.g. a NaN for an
unavailable sensor or an unspecified threshold value), asserting an
alarm doesn't make much sense, so let's not do so.

Under normal circumstances this wouldn't have arisen anyway, since the
previous behavior was to maintain the existing alarm state and the
default is deasserted, but could happen if a sensor went from available
to unavailable while a threshold alarm was asserted.