Use log crate instead of eprintln!() calls
This gives us selectable logging levels so we can retain the noisier
messages for debugging but keep them disabled by default to reduce log
spam.

The tiny manual implementation of the actual logging backend is
currently hard-coded to the INFO level.  We can replace this with a more
configurable external crate if/when needed, but figuring out which one
to use and how to configure it to not pull in a bunch of other
dependencies we don't want is more than I feel like dealing with right
now.