deps: Update gpiocdev to 0.7
Version 0.4 had a transitive dependency on a version of the 'time' crate
that was triggering security alerts (see [1]).  It's not a terribly
critical problem and I'm pretty sure unlikely to cause us any
vulnerabilities or other problems in practice, but the new gpiocdev
trims its dependencies to avoid the problem entirely and is hence nicely
a bit lighter-weight, so updating seems like a win-win.

[1] https://github.com/time-rs/time/issues/293